Such update, many pixel, very color, wow
This is a huge update. When I found that this is way more interesting
when it's at high res, I added in a ton of new features. It can now
export the images to a file, and you can use the processing moviemaker
to create (gigantic) movies of the animations. There are also some basic
keyboard functions that allow you to play around with the settings while
it runs. I've been having a blast playing with this thing.