Screen Saver Mode Activated
Due to popular demand (one of my friends), I've added in a "screensaver"
mode. Basically it just goes into full screen, as sets your grid size to
the screen size divided by the block size.

Also added in the ability to start/stop auto recording, and to be able
to toggle whether the ant is shown or not.